I think it simulates?

+++ src/Components/Node.h
@@ -25,12 +25,12 @@ namespace EC
 
 		int getAxon1() { return axon1; }
 		int getAxon2() { return axon2; }
-		int getM() { return m; }
+		float getM() { return m; }
 		int getOperation() { return operation; }
 		bool getSafeInput() { return safeInput; }
 		float getValue() { return value; }
-		int getX() { return x; }
-		int getY() { return y; }
+		float getX() { return x; }
+		float getY() { return y; }
 
 		void setAxon1(int i) { axon1 = i; }
 		void setAxon2(int i) { axon2 = i; }
@@ -40,8 +40,8 @@ namespace EC
 		void setVx(float f) { vx = f; }
 		void setVy(float f) { vy = f; }
 
-		void incrementVx(int i) { vx += i; }
-		void incrementVy(int i) { vy += i; }
+		void incrementVx(float i) { vx += i; }
+		void incrementVy(float i) { vy += i; }
 		void incrementX(float f) { x += f; }
 		void incrementY(float f) { y += f; }
 	};
